+++ oedatamodel_api/validation.py
@@ -3,9 +3,10 @@ import os
 import sys
 import csv
 import logging
+import tempfile
 from frictionless import Package, validate_resource, Report, FrictionlessException
 
-from oedatamodel_api.settings import ZIP_UPLOAD_FILEPATH
+from oedatamodel_api.settings import UPLOAD_FILEPATH
 
 # Increase max cell size (in order to successfully load timeseries.series)
 # avoid error: The data source has not supported or has inconsistent contents: field larger than field limit (131072)
@@ -18,18 +19,19 @@ class DatapackageNotValid(Exception):
     """Exception is raised if datapackage is not valid"""
 
 
-def create_and_validate_datapackage(zip_file):
-    # Save and load zip datapackage (remove saved zipfile afterwards)
-    with open(ZIP_UPLOAD_FILEPATH, "wb+") as file_object:
-        file_object.write(zip_file.file.read())
-    logger.debug("Successfully unzipped datapackage to temp folder")
-    try:
-        package = Package.from_zip(ZIP_UPLOAD_FILEPATH)
-    except FrictionlessException as fe:
-        raise DatapackageNotValid(str(fe))
-    logger.debug("Successfully loaded datapackage")
-    os.remove(ZIP_UPLOAD_FILEPATH)
-    report = create_report(package)
+def create_and_validate_datapackage(datapackage_files):
+    # Save and load datapackage into temporary folder:
+    with tempfile.TemporaryDirectory(dir=UPLOAD_FILEPATH) as tempdir:
+        for upload_file in datapackage_files:
+            with open(os.path.join(tempdir, upload_file.filename), "wb+") as file_object:
+                file_object.write(upload_file.file.read())
+        logger.debug("Successfully extracted datapackage to temp folder")
+        try:
+            package = Package(f"{tempdir}/datapackage.json")
+        except FrictionlessException as fe:
+            raise DatapackageNotValid(str(fe)) from fe
+        logger.debug("Successfully loaded datapackage")
+        report = create_report(package)
     if report.valid:
         return package
     else: